+++ src/main/java/roomescape/web/controller/AdminTimeController.java
@@ -6,8 +6,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.domain.reservation.ReservationTime;
+import roomescape.domain.user.User;
+import roomescape.service.AuthService;
 import roomescape.service.ReservationTimeCommandService;
 import roomescape.service.ReservationTimeQueryService;
+import roomescape.web.common.LoginUser;
 import roomescape.web.dto.request.ReservationTimeRequest;
 import roomescape.web.dto.response.ReservationTimeResponse;
 
@@ -21,9 +24,13 @@ public class AdminTimeController {
 
     private final ReservationTimeCommandService reservationTimeCommandService;
     private final ReservationTimeQueryService reservationTimeQueryService;
+    private final AuthService authService;
 
     @GetMapping
-    public ResponseEntity<List<ReservationTimeResponse>> getAllTimes() {
+    public ResponseEntity<List<ReservationTimeResponse>> getAllTimes(
+            @LoginUser User user
+    ) {
+        authService.validateAdmin(user);
         List<ReservationTime> allReservationTimes = reservationTimeQueryService.findAllReservationTimes();
         List<ReservationTimeResponse> reservationTimeResponse = allReservationTimes.stream()
                 .map(ReservationTimeResponse::from)
@@ -33,8 +40,10 @@ public class AdminTimeController {
 
     @PostMapping
     public ResponseEntity<ReservationTimeResponse> createReservationTime(
+            @LoginUser User user,
             @Valid @RequestBody ReservationTimeRequest request
     ) {
+        authService.validateAdmin(user);
         ReservationTime reservationTime = reservationTimeCommandService.create(request.startAt());
         Long savedId = reservationTime.getId();
 
@@ -49,8 +58,10 @@ public class AdminTimeController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservationTime(
+            @LoginUser User user,
             @PathVariable Long id
     ) {
+        authService.validateAdmin(user);
         reservationTimeCommandService.delete(id);
         return ResponseEntity.noContent().build();
     }
